Move arg parsing for train.py to ai_io.py and add options.py.
